refactor(cli): move run command to its own module
